Progres 19.05
Masih error cors
