+++ backend/src/server.js
@@ -10,7 +10,11 @@ const userRoutes = require("./routes/user");
 const saveMenuRoutes = require("./routes/save-menu");
 
 const prisma = new PrismaClient();
-const FRONTEND_URL = "https://nutrikids-v2-production-a1b6.up.railway.app";
+const FRONTEND_URL = [
+  "http://localhost:8110",
+  "https://nutrikids-v2-production-a1b6.up.railway.app"
+];
+
 
 const init = async () => {
   const server = Hapi.server({
@@ -20,7 +24,7 @@ const init = async () => {
       cors: {
         origin: [FRONTEND_URL],
         credentials: true,
-        additionalHeaders: ['cache-control', 'x-requested-with'],
+        additionalHeaders: ['cache-control', 'x-requested-with', 'content-type', 'authorization'],
         additionalExposedHeaders: ['authorization'],
       },
     },
@@ -40,29 +44,13 @@ const init = async () => {
   });
 
    server.ext("onPreResponse", (request, h) => {
-    const response = request.response;
+  if (request.method === "options") {
+    return h.response().code(200);
+  }
+  return h.continue;
+});
 
-    if (request.method === "options") {
-      return h
-        .response()
-        .code(200)
-        .header("Access-Control-Allow-Origin", FRONTEND_URL)
-        .header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-        .header("Access-Control-Allow-Headers", "Content-Type, Authorization");
-    }
 
-    if (response.isBoom) {
-      response.output.headers["Access-Control-Allow-Origin"] = "*";
-      response.output.headers["Access-Control-Allow-Methods"] = "GET, POST, PUT, DELETE, OPTIONS";
-      response.output.headers["Access-Control-Allow-Headers"] = "Content-Type, Authorization";
-    } else if (response.header) {
-      response.header("Access-Control-Allow-Origin", "*");
-      response.header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS");
-      response.header("Access-Control-Allow-Headers", "Content-Type, Authorization");
-    }
-
-    return h.continue;
-  });
 
   server.route(commentRoutes);
   server.route(authRoutes);
