+++ src/api/index.js
@@ -4,7 +4,7 @@ const isLocalhost = window.location.hostname === "localhost";
 
 const BACKEND_URL = isLocalhost
   ? "http://localhost:3000"
-  : "https://nutrikids-backend-production.up.railway.app";
+  : "nutrikids-v2-production-a1c4.up.railway.app";
 
 const ML_URL = isLocalhost
   ? "http://localhost:8000"
