+++ webpack.dev.js
@@ -9,7 +9,7 @@ module.exports = merge(common, {
     allowedHosts: 'all',
     host: '0.0.0.0',  
     static: "./dist",
-    historyApiFallback: true, // penting untuk SPA routing
+    historyApiFallback: true, 
     port: 8110,
     open: true,
     hot: true,
